backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-multipart==0.0.9
redis==5.0.4
fastapi-limiter==0.1.6
Pillow==10.3.0
imagehash==4.3.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
